Only use one todos sync saga at a time.
