+++ example/src/redux/sagas/index.js
@@ -1,19 +1,19 @@
+import { fork } from 'redux-saga/effects'
+
 import analytics from './analytics'
-import firestore from './firestore'
 import functions from './functions'
 import login from './login'
 import messaging from './messaging'
 import storage from './storage'
-// import todos from './todos'
+import todos from './todos'
 
 export default function * rootSaga () {
   yield [
-    analytics(),
-    firestore(),
-    functions(),
-    login(),
-    messaging(),
-    storage()
-    // todos(),
+    fork(analytics),
+    fork(functions),
+    fork(login),
+    fork(messaging),
+    fork(storage),
+    fork(todos)
   ]
 }
